Перенёс функции, отвечающие за показ ошибок, в отдельный модуль.
