+++ js/main.js
@@ -1,6 +1,6 @@
 import { drawMiniatures } from './draw-miniatures.js';
 import { getData } from './api.js';
-import { showGetDataErrorMessage } from './util.js';
+import { showGetDataErrorMessage } from './system-modal-messages.js';
 import { getPhotos } from './draw-full-images.js';
 import './upload-image-form.js';
