+++ js/upload-image-form.js
@@ -2,7 +2,7 @@ import { BUTTON_CLOSE_MODULE_WINDOW, IMAGE_DESCRIPTION_MAX_LENGTH } from './cons
 import { resetScale } from './upload-image-scale.js';
 import { initEffect, resetEffect } from './upload-image-effects.js';
 import { sendData } from './api.js';
-import { showSendDataErrorMessage, showSendDataSuccessMessage } from './util.js';
+import { showSendDataErrorMessage, showSendDataSuccessMessage } from './system-modal-messages.js';
 
 const imgUploadElement = document.querySelector('.img-upload');
 const imgUploadForm = imgUploadElement.querySelector('.img-upload__form');
